% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_exp_data.R
\name{extract_obs_var}
\alias{extract_obs_var}
\title{Extract Experimental Data by Variable from agroclimR Data Workbook}
\usage{
extract_obs_var(obs_data, variable, model = "oryza")
}
\arguments{
\item{obs_data}{A list containing one or more agroclimR data workbooks, as read by the \code{read_agroclimr_data()} function. Each element of the list should be a named list representing a single workbook, where each name-value pair corresponds to a specific type of observational data.}

\item{variable}{A character string specifying the variable to extract. Valid options are "phen" (phenological data), "lai" (leaf area index), "dry_matter" (dry matter), and "yield". This parameter determines which type of data the function will extract from the provided workbooks.}

\item{model}{A character string specifying the crop model for which data is being extracted. Valid options include "oryza" (for rice), "dssat" (for various crops), and "aquacrop" (for water-driven crop growth). This parameter allows the function to tailor the extraction process to the data structure used by different crop models.}
}
\value{
A \code{tibble} containing the extracted data for the specified variable and model. The returned tibble is structured to facilitate further analysis and visualization, making it a valuable resource for agricultural researchers and analysts.
}
\description{
Extracts experimental data based on specified variables from a set of agroclimR data workbooks. This function is tailored to handle data in the agroclimR xlsx format, facilitating the retrieval of agricultural research data for different variables and models. It supports a range of variables such as phenological data, leaf area index, dry matter, and yield, across various crop models.
}
\examples{
# Prepare a sample list of agroclimR data workbooks
obs_data = list(
  list(
    AGRO_man = agro,
    FERT_obs = fertil,
    PHEN_obs = phenol,
    PLANT_obs = plant,
    YIELD_obs = yield,
    SOIL_obs = soil,
    WTH_obs = weather
  )
)

# Extract phenological data for the "oryza" model
phenological_data <- extract_obs_var(obs_data, "phen", model = "oryza")
print(phenological_data)

# Extract yield data for the "oryza" model
yield_data <- extract_obs_var(obs_data, "yield", model = "oryza")
print(yield_data)

}
