% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertil.R
\docType{data}
\name{fertil}
\alias{fertil}
\title{Observed Fertilization Plan Data}
\format{
A data frame of 37 rows and 7 columns
\describe{
\item{ID}{Trial ID}{character}{XXXX}
\item{LOC_ID}{Locality ID}{character}{name}
\item{FERT_No}{Fertilizer application number}{numeric}{number}
\item{DDE}{Days after Emergence}{numeric}{days}
\item{N}{Nitrogen}{numeric}{kg/ha}
\item{P}{Phosphate}{numeric}{kg/ha}
\item{K}{Potassium}{numeric}{kg/ha}
}
}
\source{
FEDEARROZ-FNA
}
\usage{
fertil
}
\description{
A dataset containing observed fertilization management or application data.
}
\keyword{datasets}
