% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenol.R
\docType{data}
\name{phenol}
\alias{phenol}
\title{Observed Plant Phenological Data}
\format{
A data frame of 7 rows and 8 columns
\describe{
\item{ID}{Trial ID}{character}{XXXX}
\item{LOC_ID}{Locality ID}{character}{name}
\item{CULTIVAR}{Cultivar name}{character}{name}
\item{EDAT}{Emergence date}{date}{MM/DD/YYYY}
\item{IDAT}{Panicle initiation date}{date}{MM/DD/YYYY}
\item{FDAT}{Flowering date}{date}{MM/DD/YYYY}
\item{MDAT}{Maturity date}{date}{MM/DD/YYYY}
}
}
\source{
FEDEARROZ-FNA
}
\usage{
phenol
}
\description{
A dataset containing observed plant phenological dates.
}
\keyword{datasets}
