% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_exp_data.R
\name{read_agroclimr_data}
\alias{read_agroclimr_data}
\title{Read agroclimR Data}
\usage{
read_agroclimr_data(agroclimR_workbook)
}
\arguments{
\item{agroclimR_workbook}{String with the workbook name (full name).}
}
\value{
agroclimR list. A List of data frames for each sheet in the workbook.
}
\description{
Reads data from agroclimR workbook sheets.
}
\examples{
#' # File names vector, extension include
name_file = c("agroclimR_workbook.xlsx")

# Files directory
test_file = system.file("extdata", name_file, package = "agroclimR")

# Import data to R lists and tibble formats
agroclimr_list = read_agroclimr_data(test_file)
agroclimr_list

}
