% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSKS_cal.R
\name{SSKS_cal}
\alias{SSKS_cal}
\alias{SSKS_Brakensiek}
\alias{SSKS_Campbell}
\alias{SSKS_Cosby}
\alias{SSKS_Dane_Puck}
\alias{SSKS_Ferrer}
\alias{SSKS_Jabro}
\alias{SSKS_Puckett}
\alias{SSKS_Rawls}
\alias{SSKS_Vereecken}
\alias{SSKS_Wosten99}
\alias{SSKS_Suleiman_Ritchie}
\alias{SSKS_Saxton}
\title{Soil Pedotransfer Function (PTF) to estimate Soil Saturated Hydraulic Conductivity (SSKS)}
\usage{
SSKS_cal(
  S,
  C,
  SOM = 1.5,
  SBDM = 1.5,
  kmin = 0.1,
  kmax = 250,
  output = "bootmean"
)

SSKS_Suleiman_Ritchie(S, C, SOM, SBDM, SPOR = NULL, WCFC = NULL)

SSKS_Saxton(S, C, SOM, SBDM, WCFC = NULL, WCWP = NULL, WCST = NULL)
}
\arguments{
\item{S}{Sand content (\%) of the soil.}

\item{C}{Clay content (\%) of the soil.}

\item{SOM}{Soil Organic Matter content (\%) of the soil. Default is 1.5.}

\item{SBDM}{Soil Bulk Density (g/cm3) of the soil. Default is 1.5.}

\item{kmin}{Minimum threshold value for SSKS (mm/h). Default is 0.1.}

\item{kmax}{Maximum threshold value for SSKS (mm/h). Default is 250.}

\item{output}{Type of output to be returned. Options are 'data', 'summary', 'bootmean', 'mean', 'bootmedian', 'median', 'min', 'max', and 'sd'. Default is 'bootmean'.}

\item{SPOR}{Total Soil porosity (\%)}

\item{WCFC}{Water Content at Field Capacity (\%) - (-33kPa)}

\item{WCWP}{Water Content at Wilting Point (\%) - (-1500 kPa)}

\item{WCST}{Water Content at Saturation (\%) - (-10kPa)}
}
\value{
Returns a list or a numeric value of SSKS (mm/h) based on the 'output' parameter. If 'output' is 'data', it returns the complete dataset. If 'output' is 'summary', it returns the summary statistics of SSKS values. If 'output' is any of the other specified types, it returns the corresponding statistic.
}
\description{
This function calculates the Saturated Hydraulic Conductivity (SSKS)
based on various soil pedotransfer functions (PTFs) using soil properties as input parameters.
}
\examples{
# Function Arguments
S <- 28
C = 15.7
SOM <- 2
SBDM = 1.5

# SSKS multi PTF
SSKS_cal(S, C)
SSKS_cal(S, C, SOM, SBDM)
SSKS_cal(S, C, output = 'summary')
SSKS_cal(S = 28.3, C = 15.7, output = 'mean')

#' #For specific SSKS PTF:
SSKS_Saxton(S, C, SOM, SBDM)
SSKS_Suleiman_Ritchie(S, C, SOM, SBDM)

}
\references{
SSKS Pedotransfer Functions:
\itemize{
\item Brakensiek et al. (1984) for SSKS_Brakensiek
\item Campbell and Shiozawa (1994) for SSKS_Campbell
\item Cosby et al. (1984) for SSKS_Cosby
\item Dane and Puckett (1994) for SSKS_Dane_Puck
\item Ferrer Julia et al. (2004) for SSKS_Ferrer
\item Jabro (1992) for SSKS_Jabro
\item Puckett et al. (1985) for SSKS_Puckett
\item RAWLS (1986) for SSKS_Rawls
\item Saxton, K. E., & Rawls, W. J. (2006). for SSKS_Saxton.
\item Suleiman, A. A., & Ritchie, J. T. (2001). for SSKS_Suleiman_Ritchie
\item Vereecken et al. (1990) for SSKS_Vereecken
\item Wosten et al. (1999) for SSKS_Wosten99
}
}
