% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agroclimR.R
\docType{package}
\name{agroclimR}
\alias{agroclimR}
\alias{agroclimR-package}
\title{agroclimR: Streamlining Agroclimatic Data for Crop Modeling}
\description{
\code{agroclimR} is an R package developed from years of agricultural research and application,
designed to simplify the management and transformation of agroclimatic data into crop model-ready formats.
It supports major modeling systems like DSSAT v 4.8, ORYZA v3.0, and Aquacrop v6.0.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jrodriguez88.github.io/agroclimR/}
}

}
\author{
Jeferson Rodriguez-Espinoza \email{jrodriguezespinoza@outlook.com}
}
