% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agroclimR.R
\docType{package}
\name{agroclimR}
\alias{agroclimR}
\alias{agroclimR-package}
\title{agroclimR: A R-Toolkit for Agroclimatic Data Management and Crop Model File Creation}
\description{
agroclimR is a package designed to streamline the management,manipulation,
and generation of process-based model files for systems like DSSAT v 4.8,
ORYZA v3.0, and Aquacrop v6.0.
}
\author{
Jeferson Rodriguez \email{jrodriguezespinoza@outlook.com}
}
