% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETo_cal.R
\name{ETo_cal}
\alias{ETo_cal}
\title{Write AquaCrop Weather File}
\usage{
ETo_cal(wth_data, lat, elev, ref_ht = 2, kRs = 0.175, ws_mean = 2)
}
\arguments{
\item{wth_data}{A Data frame Weather data. minimum = date, tmax, tmin, rain}

\item{elev}{Numeric. Elevation (meters above sea level)}

\item{kRs}{Numeric. Hargreaves empirical coefficient, Coastal coefficient}

\item{ws_mean}{Numeric. Wind Speed Mean, default = 2 m/s}

\item{lon}{Numeric. Longitude (decimal degrees)}
}
\value{
This function returns a \code{Numeric} ET0 values (mm).
}
\description{
Function compute weather information AquaCrop v6.0  weather file.
}
\examples{
# Write file
ETo_cal(wth_data = weather, lat = 3.8, elev = 650)

}
