% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_STC.R
\name{get_STC}
\alias{get_STC}
\title{Calculate Soil Texture Class}
\usage{
get_STC(S, C, sysclass = "USDA")
}
\arguments{
\item{S}{Numeric vector indicating the percentage of sand in the soil.}

\item{C}{Numeric vector indicating the percentage of clay in the soil.}

\item{sysclass}{Character string specifying the soil classification system to use.
Default is "USDA", but can be changed if other systems are supported by the \code{soiltexture} package.}
}
\value{
String with the calculated soil texture class name, based on the input percentages of sand, clay, and silt.
}
\description{
This function calculates the soil texture class based on the percentages of sand and clay
using the specified soil classification system (default is USDA).
}
\note{
This function requires the \code{soiltexture} package to calculate the soil texture class.
Ensure that the \code{soiltexture} package is installed and loaded into your R session.
}
\examples{
# Calculate the soil texture class for a soil with 30\% sand and 20\% clay using USDA system
get_STC(30, 20)
}
\references{
Soil Texture, The Soil Texture Wizard: \if{html}{\out{<URL to soiltexture or related documentation>}}
}
