% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_exp_data.R
\name{import_exp_data}
\alias{import_exp_data}
\title{Import Experimental data from agroclimR_data workbook}
\usage{
import_exp_data(path = ., files, model = "oryza")
}
\arguments{
\item{path}{A string indicating path folder or working directory}

\item{files}{A String vector. Vector of strings from workbooks names}

\item{model}{A string of workbook names}
}
\value{
This function returns a \code{list} if files created in path folder.
}
\description{
Function to extract experimental data from agroclimR_data workbook
}
\examples{
# Import Experimental data from agroclimR_data workbook
path = "data/"
files <- list.files(path = "data/", pattern = "xls")
import_exp_data <- function(path = ., files, model = "oryza")

}
