% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_exp_data.R
\name{import_exp_data}
\alias{import_exp_data}
\title{Import Experimental Data from agroclimR Data Workbook}
\usage{
import_exp_data(files, model = "oryza")
}
\arguments{
\item{files}{Vector of strings with workbook names (.xls extension file included).}

\item{model}{String indicating the model name, default "oryza",options: ("dssat", "aquacrop").}
}
\value{
List containing raw and observed data by component ("data", "soil", "wth", "phen", "lai", "dry_matter", "yield").
}
\description{
\code{import_exp_data()} performs extraction of experimental data from the agroclimR data workbook, including
location, soil, climate data, and observed data for phenology, leaf area index,
dry matter, and yield.
}
\examples{
#' # File names vector, extension include
name_file = c("SDTO_F2000.xlsx")

# Files directory
test_file = system.file("extdata", name_file[1], package = "agroclimR")

# Import data to R lists and tibble formats
obs_data = import_exp_data(test_file, model = "oryza")

head(obs_data)
}
\seealso{
\link[sirad]{?sirad}
}
