% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant.R
\docType{data}
\name{plant}
\alias{plant}
\title{Observed Plant Growth Data}
\format{
A data frame of 77 rows and 22 columns
\describe{
\item{ID}{Trial ID}{character}{XXXX}
\item{LOC_ID}{Locality ID}{character}{name}
\item{PROJECT}{Project ID}{character}{name}
\item{CULTIVAR}{Cultivar name}{character}{name}
\item{SAMPLING_DATE}{Sampling date}{date}{MM/DD/YYYY}
\item{LAI_OBS}{Leaf Area Index observed}{numeric}{m² leaf/m² ground}
\item{LAI_SE}{Leaf Area Index standard deviation}{numeric}{m² leaf/m² ground}
\item{WLVG_OBS}{Green leaf dry weight}{numeric}{kg/ha}
\item{WLVG_SE}{Green leaf dry weight standard error}{numeric}{kg/ha}
\item{WLVD_OBS}{Dead leaf dry weight}{numeric}{kg/ha}
\item{WLVD_SE}{Dead leaf dry weight standard error}{numeric}{kg/ha}
\item{WST_OBS}{Stem dry weight}{numeric}{kg/ha}
\item{WST_SE}{Stem dry weight standard error}{numeric}{kg/ha}
\item{WSO_OBS}{Panicle dry weight}{numeric}{kg/ha}
\item{WSO_SE}{Panicle dry weight standard error}{numeric}{kg/ha}
\item{WAGT_OBS}{Total dry weight}{numeric}{kg/ha}
\item{WAGT_SE}{Total dry weight standard error}{numeric}{kg/ha}
\item{NLV_OBS}{Number of green leaves}{numeric}{number/m²}
\item{NLV_SE}{Number of green leaves standard error}{numeric}{number/m²}
\item{NST_OBS}{Number of stems}{numeric}{number/m²}
\item{NST_SE}{Number of stems standard error}{numeric}{number/m²}
\item{NP_OBS}{Number of panicles}{numeric}{number/m²}
\item{NP_SE}{Number of panicles standard error}{numeric}{number/m²}
}
}
\source{
FEDEARROZ-FNA
}
\usage{
plant
}
\description{
A dataset containing observed plant growth data (organ biomass or dry matter samplings).
}
\keyword{datasets}
