% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\docType{data}
\name{soil}
\alias{soil}
\title{Observed Soil Data}
\format{
A data frame of 18 rows and 22 columns
\describe{
\item{NL}{Number of soil layers (maximum is 10)}{numeric}{number}
\item{DEPTH}{Thickness of each soil layer}{numeric}{cm}
\item{STC}{Soil Texture Class (12-USDA)}{character}{USDA texture class}
\item{SAND}{Soil sand content}{numeric}{\%}
\item{SILT}{Soil silt content}{numeric}{\%}
\item{CLAY}{Soil clay content}{numeric}{\%}
\item{SBDM}{Soil Bulk Density}{numeric}{g/cm³}
\item{SOC}{Soil organic carbon}{numeric}{g/kg}
\item{SLON}{Soil Organic Nitrogen}{numeric}{mg/kg}
\item{SNH4}{Ammonium, KCl, g elemental N}{numeric}{mg/kg}
\item{SNO3}{Nitrate, KCl, g elemental N}{numeric}{mg/kg}
\item{PH}{pH}{numeric}{number}
\item{SCEC}{Cation exchange capacity}{numeric}{cmol/kg}
\item{WCST}{Saturated volumetric water content}{numeric}{\%}
\item{WCFC}{Volumetric water content at field capacity}{numeric}{\%}
\item{WCWP}{Volumetric water content at wilting point}{numeric}{\%}
\item{WCAD}{Volumetric water content at air dryness}{numeric}{\%}
\item{SSKS}{Saturated hydraulic conductivity}{numeric}{cm/h}
}
}
\source{
FEDEARROZ-FNA
}
\usage{
soil
}
\description{
A dataset containing observed soil data.
}
\keyword{datasets}
