% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_exp_oryza.R
\name{write_exp_oryza}
\alias{write_exp_oryza}
\title{Write ORYZA v3 Experimental File (.EXP)}
\usage{
write_exp_oryza(agroclimr_list, path, ET_mod = "PRIESTLY TAYLOR")
}
\arguments{
\item{agroclimr_list}{R list imported from excel workbook using \code{read_agroclimr_data()}.}

\item{path}{A string indicating path folder or working directory}

\item{ET_mod}{String indicating is method for evapotranspiration calculation,
'PENMAN' = Penman-based (Van Kraalingen& Stol,1996), 'PRIESTLY TAYLOR' = Priestly-Taylor ("),}
}
\value{
This function returns a \code{vector} of model files created in path folder.
}
\description{
\code{write_exp_oryza()} performs transformation from experimental data to ORYZA v3 file model format.
}
\examples{
#' # File names vector, extension include
name_file = c("agroclimR_workbook.xlsx")

# Files directory
test_file = system.file("extdata", name_file, package = "agroclimR")

# Import data to R lists and tibble formats
agroclimr_list = read_agroclimr_data(test_file)

# Write Oryza Experimental Files
exp_files_created <- write_exp_oryza(agroclimr_list, path = "./")

exp_files_created
file.remove(exp_files_created)

}
