% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_soil_dssat.R
\name{write_soil_dssat}
\alias{write_soil_dssat}
\title{Write DSSAT v4.8 Soil File}
\usage{
write_soil_dssat(
  path,
  id_name,
  soil_data,
  salb = 0.13,
  evapL = 6,
  sldr = 0.6,
  slnf = 1,
  slpf = 1,
  multi = F
)
}
\arguments{
\item{path}{A string indicating path folder or working directory}

\item{id_name}{A String. 4 letters string of locality name. (ex. "JR")}

\item{soil_data}{A Data frame. Soil data. see \code{soil}}

\item{salb}{Numeric. Albedo, fraction}

\item{evapL}{Numeric. Evaporation limit, (mm)}

\item{slnf}{Numeric. Mineralization factor, 0 to 1 scale.}

\item{slpf}{Numeric. Photosynthesis factor, 0 to 1 scale}

\item{multi}{Logical. Soil annual average temperature of the first layers}

\item{max_depth}{description}
}
\value{
This function returns a \code{logical} if files created in path folder.
}
\description{
Function compute Soil information to DSSAT v4.8 soil file.
}
\examples{
# Write DSSAT v4.8 Soil file
soil = group_by(soil, NL) \%>\% sample_n(1)
write_soil_dssat(id_name = "test_soil", soil_data = soil)

}
