% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_soil_dssat.R
\name{write_soil_dssat}
\alias{write_soil_dssat}
\title{Write DSSAT v4.8 Soil File}
\usage{
write_soil_dssat(
  path = ".",
  id_name,
  soil_data,
  salb = 0.13,
  evapL = 6,
  sldr = 0.6,
  slnf = 1,
  slpf = 1,
  multi = FALSE
)
}
\arguments{
\item{path}{A string indicating path folder or working directory}

\item{id_name}{A String. 4 letters string of locality name. (ex. "JR")}

\item{soil_data}{A Data frame. Soil data. See \code{?soil}}

\item{salb}{Numeric. Albedo, fraction}

\item{evapL}{Numeric. Evaporation limit, (mm)}

\item{slnf}{Numeric. Mineralization factor, 0 to 1 scale.}

\item{slpf}{Numeric. Photosynthesis factor, 0 to 1 scale}

\item{multi}{Logical. All soil profiles in the same file}

\item{max_depth}{description}
}
\value{
This function returns a \code{logical} if files created in path folder.
}
\description{
Function compute Soil information to DSSAT v4.8 soil file.
}
\examples{
# Write DSSAT v4.8 Soil file
soil_sample = dplyr::group_by(soil, NL) |>
dplyr::sample_n(1) |> dplyr::ungroup()
soil_files_created <- write_soil_dssat(id_name = "soil_dssat", soil_data = soil_sample)

soil_files_created
file.remove(soil_files_created)

}
