% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_soil_oryza.R
\name{write_soil_oryza}
\alias{write_soil_oryza}
\title{Write ORYZA v3 Soil File}
\usage{
write_soil_oryza(
  path = ".",
  id_name,
  soil_data,
  ZRTMS = 0.5,
  WL0I = 0,
  WCLI = "FC",
  RIWCLI = "NO",
  SATAV = 20
)
}
\arguments{
\item{path}{A string indicating path folder or working directory}

\item{id_name}{A String 4 letters string of locality name. "AIHU" = Aipe, Huila}

\item{soil_data}{A Data frame. Soil data. see \code{soil}}

\item{ZRTMS}{Numeric. Maximum rooting depth in the soil (m)}

\item{WL0I}{Numeric. Initial pounded water depth at start of simulation (mm)}

\item{WCLI}{Numeric/String. WCLI can take 3 values: Field Capacity ('FC'), 50\% of Soil Saturation ('ST50'), Fraction of water content ('0.0'- '1.0')}

\item{RIWCLI}{A String. Re-initialize switch RIWCLI is ('YES') or ('NO')}

\item{SATAV}{Numeric. Soil annual average temperature of the first layers}
}
\value{
This function returns a \code{logical} if files created in path folder.
}
\description{
Function compute Soil information to ORYZA soil file.
}
\examples{
# Write ORYZA Soil file
soil = group_by(soil, NL) \%>\% sample_n(1)
write_soil_oryza(id_name = "test_soil", soil_data = soil)

}
