% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wth_aquacrop.R
\name{write_wth_aquacrop}
\alias{write_wth_aquacrop}
\title{Write AquaCrop Weather File}
\usage{
write_wth_aquacrop(
  path,
  id_name,
  wth_data,
  lat,
  lon,
  elev,
  co2_file = "MaunaLoa.CO2",
  ...
)
}
\arguments{
\item{path}{A string indicating path folder or working directory}

\item{id_name}{A String. 4 letters string of locality name. "AIHU" = Aipe, Huila}

\item{wth_data}{A Data frame Weather data. minimum = date, tmax, tmin, rain}

\item{lat}{Numeric. Latitude (decimal degrees)}

\item{lon}{Numeric. Longitude (decimal degrees)}

\item{elev}{Numeric. Elevation (meters above sea level)}

\item{co2_file}{A String. Default is "MaunaLoa.CO2".  *.CO2 files are available in Aquacrop default DB}
}
\value{
This function returns a \code{logical} if files created in path folder.
}
\description{
Function compute weather information AquaCrop v6.0  weather file.
}
\examples{
# Write file
write_wth_aquacrop(
  path = ".", id_name = "TEST", wth_data = weather,
  lat = 3.8, lon = -76.5, elev = 650)

}
