% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wth_aquacrop.R
\name{write_wth_aquacrop}
\alias{write_wth_aquacrop}
\title{Write AquaCrop Weather File}
\usage{
write_wth_aquacrop(
  path = ".",
  id_name,
  wth_data,
  lat,
  lon,
  elev,
  co2_file = "MaunaLoa.CO2",
  ...
)
}
\arguments{
\item{path}{A string indicating the path folder or working directory where the weather files will be saved.}

\item{id_name}{A 4-letter string representing the locality name abbreviation. For example, "AIHU" stands for Aipe, Huila.}

\item{wth_data}{A data frame containing weather data with at least the following columns: date, tmax, tmin, rain.}

\item{lat}{Numeric. Latitude of the location in decimal degrees.}

\item{lon}{Numeric. Longitude of the location in decimal degrees.}

\item{elev}{Numeric. Elevation of the location in meters above sea level.}

\item{co2_file}{A string representing the CO2 file to be used. Default is "MaunaLoa.CO2". CO2 files are available in the Aquacrop default database.}
}
\value{
This function returns a logical value indicating whether the files were successfully created in the specified path folder.
}
\description{
Function to compute weather information AquaCrop v6.0 weather file.
}
\examples{
# Write AquaCrop weather file
write_wth_aquacrop(
  path = ".", id_name = "TEST", wth_data = weather,
  lat = 3.8, lon = -76.5, elev = 650)

}
