% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wth_dssat.R
\name{write_wth_dssat}
\alias{write_wth_dssat}
\title{Write DSSAT v4.8 Weather File}
\usage{
write_wth_dssat(path = ".", id_name, wth_data, lat, lon, elev, ref_ht = 2)
}
\arguments{
\item{path}{A string indicating path folder or working directory}

\item{id_name}{A String 4 letters string of locality name. "AIHU" = Aipe, Huila}

\item{wth_data}{A Data frame Weather data. minimum = date, tmax, tmin, rain}

\item{lat}{Numeric. Latitude (decimal degrees)}

\item{lon}{Numeric. Longitude (decimal degrees)}

\item{elev}{Numeric. Elevation (meters above sea level)}

\item{ref_ht}{Numeric. Reference height (meters)- weather station}
}
\value{
This function returns a vector of model files created in path folder.
}
\description{
Function compute weather information DSSAT v4.8 weather file.
}
\examples{
# Write file
wth_files_created <- write_wth_dssat(
  path = ".", id_name = "TEST", wth_data = weather,
  lat = 3.91, lon = -75.0, elev = 450)

readLines(wth_files_created[1], n = 15) |> writeLines()
file.remove(wth_files_created)

}
\references{
\itemize{
\item DSSAT Weather module: \url{https://dssat.net/weather-module/}
\item Brakensiek et al. (1984) for SSKS_Brakensiek
}
}
