% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wth_oryza.R
\name{write_wth_oryza}
\alias{write_wth_oryza}
\title{Write ORYZA v3 Weather File}
\usage{
write_wth_oryza(
  path = ".",
  id_name,
  wth_data,
  lat,
  lon,
  elev,
  stn = 1,
  multiyear = F,
  tag = F
)
}
\arguments{
\item{path}{A string indicating path folder or working directory}

\item{id_name}{A String 4 letters string of locality name. "AIHU" = Aipe, Huila}

\item{wth_data}{A Data frame Weather data. minimum = date, tmax, tmin, rain}

\item{lat}{Numeric. Latitude (decimal degrees)}

\item{lon}{Numeric. Longitude (decimal degrees)}

\item{elev}{Numeric. Elevation (meters above sea level)}

\item{stn}{Integer. Station number}

\item{multiyear}{A Logical. TRUE = ".cli" multiyear format or FALSE = yearly format (ie. 1998 = *.998)}

\item{tag}{A Logical. TRUE = write information for each file}
}
\value{
This function returns a \code{logical} if files created in path folder.
}
\description{
Function compute weather information to ORYZA weather file.
}
\examples{
# Write file
write_wth_oryza(
  path = ".", id_name = "TEST", wth_data = weather,
  lat = 3.8, lon = -76.5, elev = 650)

write_wth_oryza(
  path = ".", id_name = "TEST2", wth_data = weather,
  lat = 3.8, lon = -76.5, elev = 650, multiyear = TRUE, tag = TRUE)

}
